+++ src/Command/DeployContainerHooksCommand.php
@@ -16,21 +16,17 @@ use Torr\Hosting\Deployment\TaskCli;
 		"hosting:hook:deploy",
 	],
 )]
-final class DeployContainerHooksCommand extends Command
+final readonly class DeployContainerHooksCommand
 {
 	/**
 	 */
 	public function __construct (
-		private readonly HookRunners $runners,
-	)
-	{
-		parent::__construct();
-	}
+		private HookRunners $runners,
+	) {}
 
 	/**
 	 */
-	#[\Override]
-	protected function execute (InputInterface $input, OutputInterface $output) : int
+	public function __invoke (InputInterface $input, OutputInterface $output) : int
 	{
 		$io = new TaskCli($input, $output);
 		$io->title("Run Deploy Container Hooks");
@@ -48,6 +44,6 @@ final class DeployContainerHooksCommand extends Command
 		$io->newLine();
 		$io->success("Ran all deploy container hooks.");
 
-		return 0;
+		return Command::SUCCESS;
 	}
 }