+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * (improvement) Rename `DeployHookInterface` to `DeployContainerHookInterface`.
 * (deprecation) Deprecate `DeployHookInterface`.
 * (internal) Add tests for hook management.
+* (improvement) Refactor all commands to new invokable commands.
+* (internal) Add tests for command registration + execution.
 
 
 4.2.0